Stage 제작
다음 스테이지로 넘어가는 로직 구현
Map Generate를 통해 맵 생성 알고리즘 구현

[문제점]
멀티 씬으로 하려고 하였으나 씬끼리의 물리 영역은 종속되어 다른 씬과 상호작용을 하지 않아 로드 된 씬에 있는 모든 오브젝트를 메인 씬으로 옮겨줌

+++ Bearlike Project/Assets/Scripts/State/GameStatus/Editor/StatusValueEditor.cs
@@ -45,7 +45,7 @@ namespace Scripts.State.GameStatus.Editor
 
             int sliderInterval = 150;
             var currentSliderPos = new Rect(position.x, position.y, sliderInterval, position.height);
-            current.floatValue = (int)GUI.HorizontalSlider(currentSliderPos, current.floatValue, min.floatValue, max.intValue);
+            current.floatValue = (int)GUI.HorizontalSlider(currentSliderPos, current.floatValue, min.floatValue, max.floatValue);
             sumInterval += sliderInterval + 10;
                 
             // TO DO
